// dpr/types/api.d.ts
export interface components {
  schemas: {
// ...
    DashboardDefinitionSummary: {
      id: string
      name: string
      description?: string
      /** @enum {string} */
      loadType?: 'sync' | 'async'
    }
    ReportDefinitionSummary: {
      id: string
      name: string
      description?: string
      variants: components['schemas']['VariantDefinitionSummary'][]
      dashboards?: components['schemas']['DashboardDefinitionSummary'][]
      authorised: boolean
    }
    VariantDefinitionSummary: {
      id: string
      name: string
      description?: string
      isMissing: boolean
      /** @enum {string} */
      loadType?: 'sync' | 'async'
    }
// ...
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}